pyproject.toml
[tool.poetry]
name = "pylocker"
version = "0.0.1"
description = "file storing service"
authors = ["pyux"]

[tool.poetry.dependencies]
python = "^3.8"
alembic = "^1.4.3"
fastapi = "^0.63.0"
pydantic = { extras = ["email"], version = "^1.7.3" }
python-multipart = "^0.0.5"
databases = "^0.4.1"
gunicorn = "^20.0.4"
uvicorn = "^0.13.1"
psycopg2 = "^2.8.6"
asyncpg = "^0.21.0"
config = "^0.5.0"
fusionauth-client = "^1.22.2"
loguru = "^0.5.3"
aredis = "^1.1.8"

[tool.poetry.dev-dependencies]
sqlalchemy-stubs = "^0.3"
flakehell = "^0.7.1"
mypy = { extras = ["markdown"], version = "^0.790" }
flake8 = "^3.8.4"
pydantic = { extras = ["black"], version = "^1.7.3" }
black = "^20.8b1"
python-dotenv = "^0.15.0"
pytest = "^6.2.1"
pytest-cov = "^2.10.1"
pytest-sugar = "^0.9.4"
pytest-mock = "^3.4.0"
pytest-asyncio = "^0.14.0"
pytest-dotenv = "^0.5.2"
assertpy = "^1.1"
httpx = "^0.16.1"
Faker = "^5.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.flakehell]
base = "https://raw.githubusercontent.com/life4/flakehell/master/pyproject.toml"
exclude = [".certs", ".env", ".gitignore", "Dockerfile", "docker-compose.yaml", "mypy.ini"]
format = "grouped"
max_line_length = 120
show_source = true

[tool.flakehell.plugins]
pyflakes = ["**", "-F401"]
flake8-bandit = ["-*", "+S1??"]
"flake8-*" = ["**"]
flake8-docstrings = ["-*"]
